Move GameDescription to o.t.lobby.common
