+++ game-core/src/main/java/games/strategy/engine/framework/startup/ui/InGameLobbyWatcherWrapper.java
@@ -1,8 +1,9 @@
 package games.strategy.engine.framework.startup.ui;
 
+import org.triplea.lobby.common.GameDescription.GameStatus;
+
 import games.strategy.engine.framework.IGame;
 import games.strategy.engine.framework.startup.mc.GameSelectorModel;
-import games.strategy.engine.lobby.server.GameDescription.GameStatus;
 
 /**
  * A proxy for an {@link InGameLobbyWatcher} to accommodate dynamically changing the underlying lobby watcher as games
