+++ game-core/src/main/java/games/strategy/engine/lobby/client/ui/LobbyGamePanel.java
@@ -23,13 +23,13 @@ import javax.swing.ListSelectionModel;
 import javax.swing.SwingUtilities;
 
 import org.mindrot.jbcrypt.BCrypt;
+import org.triplea.lobby.common.GameDescription;
 import org.triplea.lobby.common.IModeratorController;
 import org.triplea.lobby.common.LobbyConstants;
 import org.triplea.swing.SwingAction;
 
 import games.strategy.engine.framework.GameRunner;
 import games.strategy.engine.framework.startup.ui.ServerOptions;
-import games.strategy.engine.lobby.server.GameDescription;
 import games.strategy.net.INode;
 import games.strategy.net.Messengers;
 import games.strategy.net.Node;
